Add line break converter
